backend/requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.17.0
gunicorn==25.3.0
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.12.0
python-dotenv==1.2.2
request==2.33.1 